implemented get result with userid

+++ src/main/java/com/tm/evaluationsystem/dto/ResultResponseDTO.java
@@ -19,6 +19,6 @@ public class ResultResponseDTO {
     private List<QuestionAnswerMap> questionAnswerMaps;
     private int attemptedQuestions;
     private int rightAnswers;
-    private int percentage;
+    private float percentage;
     private boolean isPassed;
 }
